refactor: pagination response type 수정
