+++ apps/community/src/types/api.ts
@@ -1,12 +1,18 @@
-import type { Board, Pagable } from './board';
-
 type BaseResponse<T> = {
   data: T;
 };
 
-interface PaginationResponse {
-  contents: Board[];
+type PaginationResponse<T> = BaseResponse<{
+  contents: T[];
   pagable: Pagable;
+}>;
+
+interface Pagable {
+  page: number;
+  size: number;
+  totalPage: number;
+  totalElements: number;
+  isEnd: boolean;
 }
 
 export type { BaseResponse, PaginationResponse };
